tweaks

+++ libs/newt-common/src/filesystem.rs
@@ -22,6 +22,7 @@ pub struct ListFilesOptions {
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "snake_case")]
 pub enum UserGroup {
     Name(String),
     Id(u32),
@@ -77,9 +78,9 @@ pub struct FsStats {
 impl From<nix::sys::statvfs::Statvfs> for FsStats {
     fn from(stats: nix::sys::statvfs::Statvfs) -> Self {
         Self {
-            free_bytes: stats.blocks_available() * stats.fragment_size(),
-            available_bytes: stats.blocks_available() * stats.fragment_size(),
-            total_bytes: stats.blocks() * stats.fragment_size(),
+            free_bytes: ((stats.blocks_available() as u64) * (stats.fragment_size() as u64)),
+            available_bytes: ((stats.blocks_available() as u64) * (stats.fragment_size() as u64)),
+            total_bytes: ((stats.blocks() as u64) * (stats.fragment_size() as u64)),
         }
     }
 }
